--- a/markovify/utils.py
+++ b/markovify/utils.py
@@ -43,8 +43,10 @@ def combine(models, weights=None):
     if isinstance(ret_inst, Chain):
         return Chain.from_json(c)
     if isinstance(ret_inst, Text):
-        combined_text = "\n".join(m.input_text for m in models)
-        return ret_inst.from_chain(c, corpus=combined_text)
+        combined_sentences = []
+        for m in models:
+            combined_sentences += m.parsed_sentences
+        return Text(input_text=None, parsed_sentences=combined_sentences)
     if isinstance(ret_inst, list):
         return list(c.items())
     if isinstance(ret_inst, dict):