Added `runs` to the exported model to save run-time when a Text model is instantiated from an existing JSON/dict. With the current markovify there isn't really much time saved by loading an exported model, but by saving `runs` the time is more than halved.

Please let me know what you think.

Benchmarking method (see [gist](https://gist.github.com/ammgws/9d2b4d278ae281d3ef282c437c0cef5e))
a) Create new model from scratch `markovify.Text(corpus, state_size=2, chain=None)`
b) Load an existing chain JSON `markovify.Text.from_chain(chain_json, corpus=corpus)`
c) Load an existing model JSON `markovify.Text.from_json(model_json)`

**Current markovify**
- a) RasPi 16.2 secs, Desktop 0.24 secs
- b) RasPi 14.1 secs, Desktop 0.22 secs
- c) RasPi 15.4 secs, Desktop 0.25 secs

**This branch**
- a) RasPi 16.3 secs, Desktop 0.23 secs
- b) RasPi 13.7 secs, Desktop 0.21 secs
- c) RasPi 06.3 secs, Desktop 0.10 secs

RasPi = Raspberry Pi 2
Desktop = Intel i5-3570K CPU @ 3.40GHz, 8GB RAM