@jsvine 2017-01-05
Thanks, Juan! I like this idea in general. Two thoughts re. tweaks:

- I'd prefer if this feature were opt-in. Including the runs increases the size of the exported JSON/dict, and some users might not want to do that.

- Let's think about other variable names we could use instead of `runs`. That name is somewhat semantically ambiguous. (I should have chosen a more descriptive name when I initially wrote this library.) Maybe `sentence_lists` or `parsed_text`? Open to other ideas, too.