@ammgws 2017-01-05
Regarding the size of the exported model, what I really wanted to do was
remove input_text from the export and in place of it export either
rejoined_text or runs since IIRC input_text is only used to generate the
other two and isn't referenced by any other functions in the class. That
way the file size should be the same as it is now or even smaller (in the
case of rejoined_text since it removes certain characters etc).

However that would break compatibility with models exported using older
versions so I left it going to discuss here.

On 5 Jan 2017 9:33 am, "Jeremy Singer-Vine" <notifications@github.com>
wrote:

> Thanks, Juan! I like this idea in general. Two thoughts re. tweaks:
>
>    -
>
>    I'd prefer if this feature were opt-in. Including the runs increases
>    the size of the exported JSON/dict, and some users might not want to do
>    that.
>    -
>
>    Let's think about other variable names we could use instead of runs.
>    That name is somewhat semantically ambiguous. (I should have chosen a more
>    descriptive name when I initially wrote this library.) Maybe
>    sentence_lists or parsed_text? Open to other ideas, too.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/jsvine/markovify/pull/50#issuecomment-270528413>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/ATc742VN0kv-WmmQMRliYxxckzJoiY54ks5rPDpzgaJpZM4LaYmE>
> .
>