@ekt1701 2017-01-21
Hello, I tried to test this pull request but I'm getting this error message:

`File "/usr/local/lib/python2.7/site-packages/markovify/text.py", line 40
    "runs": self.runs
SyntaxError: invalid syntax`

I simply replaced the existing text.py with this version.