@ammgws 2017-01-30
Yeah you will need to regenerate the JSON file using the updated code, since it adds extra data to the exported JSON file.