@ekt1701 2017-01-31
I tried to make a new JSON file with the updated code, however, I got this error:

```
File "makej.py", line 1, in <module>
    import markovify
  File "/usr/local/lib/python2.7/site-packages/markovify/__init__.py", line 5, in <module>
    from .text import Text, NewlineText
  File "/usr/local/lib/python2.7/site-packages/markovify/text.py", line 40
    "runs": self.runs
SyntaxError: invalid syntax
```