@ekt1701 2017-02-01
Thank you, the syntax errors are gone.  Hopefully, tomorrow, I can test the speed difference.