@jsvine 2017-02-07
Thanks, @ammgws and thanks @ekt1701. Just checking: How did the speed compare?