@ekt1701 2017-02-10
Here are my results with a text file consisting of 36,166 quotes.

Running 25 times each, returning the average runtime in seconds.

Current markovify, create new model, state size 2:
3.36153388023
Example: But chemical agriculture offered new tools that are different, but they all met and became Picasso.
----------------------------------
Current markovify, existing chain JSON, state size 2:
3.43817088127
Example: I want to see that in the Ann Arbor area.
----------------------------------
Current markovify, existing model JSON, state size 2:
4.69978396416
Example: Love and kindness from the challenges and adversity.
----------------------------------

Modified markovify, create new model, state size 2:
3.41913988113
Example: I heard dogs bark more fiercely than usual, I stood against the people who you trusted?
----------------------------------
Modified markovify, existing chain JSON, state size 2:
3.2639552784
Example: I say that, of course.
----------------------------------
Modified markovify, existing model JSON, state size 2:
2.6922947216
Example: History provides neither compensation for suffering is transformed into an extraordinary adventure.
----------------------------------