@Abhishaike 2017-02-14
Hi ammgws, is there a way to extend this idea of using 'runs' to the POSified method of generating markov chains? I've attached the POSified class below. 

```
class POSifiedText(markovify.Text): 
    def word_split(self, sentence):
        words = re.split(self.word_split_pattern, sentence)
        words = [ "::".join(tag) for tag in nltk.pos_tag(words) ]
        return words
    def word_join(self, words):
        sentence = " ".join(word.split("::")[0] for word in words)
        return sentence
```