@ammgws 2017-02-15
All the POSified method does is (from were I can see) is override the
functions for how the corpus is processed. The idea of 'runs' is still the
same as it subclasses markovify, so I don't see why it wouldn't work. Did
you give it a go?

On 15 Feb. 2017 03:40, "Abhishaike" <notifications@github.com> wrote:

Hi ammgws, is there a way to extend this idea of using 'runs' to the
POSified method of generating markov chains? I've attached the POSified
class below.

class POSifiedText(markovify.Text):
    def word_split(self, sentence):
        words = re.split(self.word_split_pattern, sentence)
        words = [ "::".join(tag) for tag in nltk.pos_tag(words) ]
        return words
    def word_join(self, words):
        sentence = " ".join(word.split("::")[0] for word in words)
        return sentence

—
You are receiving this because you were mentioned.
Reply to this email directly, view it on GitHub
<https://github.com/jsvine/markovify/pull/50#issuecomment-279795903>, or mute
the thread
<https://github.com/notifications/unsubscribe-auth/ATc747E8BeMrMS8ibG5kXAtSd0YKKQFoks5rcfUtgaJpZM4LaYmE>
.