@jsvine 2017-03-23
I'm looking to incorporate these suggestions, but it seems this PR is failing a test:

```
$ nosetests
.........F
======================================================================
FAIL: test_simple (test.test_combine.MarkovifyTest)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/Users/jeremysinger-vine/Dropbox/personal/projects/markovify/test/test_combine.py", line 18, in test_simple
    assert(combo.chain.model == text_model.chain.model)
AssertionError
```

Here's the relevant test:

```python
class MarkovifyTest(unittest.TestCase):

    def setUp(self):
        with open(os.path.join(os.path.dirname(__file__), "texts/sherlock.txt")) as f:
            self.sherlock = f.read()

    def test_simple(self):
        text_model = markovify.Text(self.sherlock)
        combo = markovify.combine([ text_model, text_model ], [ 0.5, 0.5 ])
        assert(combo.chain.model == text_model.chain.model)
```

Do you know what might be causing the failure? It shouldn't be affected by changes to model-exporting.